% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum.N.R
\name{cum.N}
\alias{cum.N}
\title{Calculate the number of accrued samples (Not for all)}
\usage{
cum.N(Sa, ra.seq)
}
\arguments{
\item{Sa}{Accrual duration}

\item{ra.seq}{Sequence of piecewise accrual rate. If length(ra.seq)>Sa, it directly assumes using ra.seq[-1] (the last one) as the accrual rate for the rest accrual period}
}
\value{
A number of accrued samples
}
\description{
Function used for calculating number of cumulative accrued samples based on a piecewise accrual rate. Used for piecewise accrual case.
}
\examples{
cum.N(10, c(10,15,20,30,25))
}
