% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E1.R
\name{E1}
\alias{E1}
\title{Calculate the expected number of events (single arm, uniform accrual rate)}
\usage{
E1(S = NULL, N = NULL, Sa = NULL, ra = NULL, lambda, eta)
}
\arguments{
\item{S}{study duration, or equivalently, accrual time (Sa) + follow-up time (Sf)}

\item{N}{number of subjects}

\item{Sa}{accrual duration, a numeric value}

\item{ra}{uniform accural rate, a numeric value}

\item{lambda}{rate parameter for event time (assuming an exponential distribution), a numeric value > 0}

\item{eta}{rate parameter for loss to follow-up (assuming an exponential distribution), a numeric value > 0}
}
\value{
Number of exepected events under such trial design settings
}
\description{
A very fundamental function to calculate the expected number of events given sample size (N), study duration (S), and either one of accrual rate (ra) or accural duration (Sa)
}
\details{
Only need three of S, N, ra, and Sa. Keep unit the same among these variables
}
\examples{
E1(S = 24, N = 120, ra = 10, lambda = log(2)/20, eta = 0)
}
\references{
Kim, Kyungmann, and Anastasios A. Tsiatis. ``Study duration for clinical trials with survival response and early stopping rule." Biometrics (1990): 81-92.
}
