% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_a.R
\name{E_a}
\alias{E_a}
\title{Calculate the desired number of events at the end of study}
\usage{
E_a(HR, alpha, power, alloc)
}
\arguments{
\item{HR}{Hazard Ratio, a numeric value in (0, 1)}

\item{alpha}{type I error of one-sided log-rank test, a numeric value in (0, 1)}

\item{power}{power required to observe the significant result under current settings, a numeric value in (0, 1)}

\item{alloc}{allocation between two arms, i.e. arm 1/(arm1 + arm2), a numeric value in (0,1)}
}
\value{
Number of exepected events, an integer value (rounding-up)
}
\description{
A function to compute the number of events based on asymptotic theory of log-rank test
}
\examples{
E_a(HR=0.5, alpha=0.025, power=0.9, alloc=0.5)
}
\references{
Schoenfeld, David. "The asymptotic properties of nonparametric tests for comparing survival distributions." Biometrika 68.1 (1981): 316-319.
}
