% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_piece1.R
\name{E_piece1}
\alias{E_piece1}
\title{Calculate the expected number of events (single arm, piecewise accrual rate)}
\usage{
E_piece1(S, Sa, r.seq, lambda, eta)
}
\arguments{
\item{S}{study duration, or equivalently, accrual time (Sa) + follow-up time (Sf)}

\item{Sa}{accrual duration, a numeric value}

\item{r.seq}{Sequence of piecewise accrual rate. If length(ra.seq)>Sa, it directly assumes using ra.seq[-1] (the last one) as the accrual rate for the rest accrual period}

\item{lambda}{rate parameter for event time (assuming an exponential distribution), a numeric value > 0}

\item{eta}{rate parameter for loss to follow-up (assuming an exponential distribution), a numeric value > 0}
}
\value{
Number of exepected events
}
\description{
A very fundamental function to calculate the expected number of events given sample size (N), study duration (S), and either one of accrual rate (ra) or accural duration (Sa)
}
\details{
Only need three of S, N, ra, and Sa. Keep unit the same among these variables
}
\examples{
E_piece1(S = 24, Sa = 12, r.seq = c(10,15,20,30,25), lambda = log(2)/20, eta = 0)
}
\references{
Kim, Kyungmann, and Anastasios A. Tsiatis. ``Study duration for clinical trials with survival response and early stopping rule." Biometrics (1990): 81-92.
}
