% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_piece2.R
\name{E_piece2}
\alias{E_piece2}
\title{Calculate the expected number of events in two-arm studies (uniform)}
\usage{
E_piece2(
  S = NULL,
  Sa = NULL,
  r.seq = NULL,
  lambda1,
  lambda2,
  eta1 = 0,
  eta2 = 0,
  alloc = 0.5,
  CEILING = FALSE
)
}
\arguments{
\item{S}{study duration, or equivalently, accrual time (Sa) + follow-up time (Sf)}

\item{Sa}{accrual duration, a numeric value}

\item{r.seq}{Sequence of piecewise accrual rate. If length(ra.seq)>Sa, it directly assumes using ra.seq[-1] (the last one) as the accrual rate for the rest accrual period}

\item{lambda1}{rate parameter for event time (which is an exponential distribution) in arm#1, a numeric value > 0}

\item{lambda2}{similar to lambda1, used for arm#2}

\item{eta1}{rate parameter for censoring time (which is an exponential distribution) in arm#1, a numeric value > 0}

\item{eta2}{similar to eta1, used for arm#2}

\item{alloc}{allocation between two arms, i.e. arm 1/(arm1 + arm2), a numeric value in (0,1)}

\item{CEILING}{if true, round-up the number of expected events}
}
\value{
the number of exepected events, could be round up by specifying CEILING = TRUE
}
\description{
A function to compute the number of events in two-arm studies under the assumption of exponential distribution for survival time and loss to follow-up time
}
\examples{
E_piece2(S = 24, Sa = 12, r.seq = c(10,15,20,30,25),
lambda1 = log(2)/20, lambda2 = log(2)/10, alloc = 0.66)
}
\references{
Kim, Kyungmann, and Anastasios A. Tsiatis. ``Study duration for clinical trials with survival response and early stopping rule." Biometrics (1990): 81-92.
}
