% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EF.R
\name{EF}
\alias{EF}
\title{Calculate the expected min(F(x)) via simulation}
\usage{
EF(S, Sa, ra = NULL, N, lambda, eta, nsim = 1000, seed = 12345)
}
\arguments{
\item{S}{study duration, or equivalently, accrual time (Sa) + follow-up time (Sf)}

\item{Sa}{accrual duration, a numeric value}

\item{ra}{either uniform or piecewise accural rate. See details for more information}

\item{N}{number of subjects}

\item{lambda}{rate parameter for event time (assuming an exponential distribution), a numeric value > 0}

\item{eta}{rate parameter for loss to follow-up (assuming an exponential distribution), a numeric value > 0}

\item{nsim}{Number of simulations}

\item{seed}{Seed for reproducibility}
}
\value{
A numeric value between (0,1), indicating the probability of observing the median survival time from KM curve. Or in other word, the last point of KM curve is under 0.5

Number of exepected events
}
\description{
Calculate the probability of observing the median survival time from KM curve using Monte Carlo simulation of Ui given S, Sa, N, lambda, eta
}
\examples{
EF(S = 24, Sa = 12, ra = c(10,15,20,30,25), N = 320, lambda = log(2)/20, eta = 0)
}
