% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOpt.R
\name{findOpt}
\alias{findOpt}
\title{Find optimal clinical trial design}
\usage{
findOpt(
  alpha = 0.05,
  POWER = 0.9,
  ra = NULL,
  Sa = NULL,
  Ea,
  lambda,
  eta,
  alloc = 0.5,
  c0,
  c1,
  c2,
  b,
  dL,
  d0_r0 = NULL,
  Aj.Ind = "NotUsed",
  t0 = NA,
  e0 = NA,
  p0 = NA,
  m0 = NA,
  ...
)
}
\arguments{
\item{alpha}{type I error rate (size)}

\item{POWER}{power (1-beta)}

\item{ra}{accrual rate (NULL if Sa is input)}

\item{Sa}{maximum enrollment period (NULL if ra is input), see details}

\item{Ea}{target event number to observe at the end of the study}

\item{lambda}{a vector of length 2, representing lambda (rate parameter of exponential distribution) for each arm (first treatment, then control)}

\item{eta}{a vector of length 2, representing eta (rate parameter of exponential distribution) for each arm (first treatment, then control)}

\item{alloc}{allocation between two arms, i.e. arm 1/(arm1 + arm2), a numeric value in (0,1)}

\item{c0}{fixed cost}

\item{c1}{cost per patient}

\item{c2}{cost per unit time}

\item{b}{revenue per unit time}

\item{dL}{LOE(loss of exclusivity time period) - l0 (time between final analysis and market access)}

\item{d0_r0}{threshold value for clinical meaningful results, corresponding to the choice of Aj.Ind (see reference paper)}

\item{Aj.Ind}{Indicator for clincial meaningful results. either "Median Difference" (A1) or "Median Ratio" (A2). If not used, no need to input}

\item{t0}{threshold value for data maturity requirements C1. Either one input will activate the corresponding constraint}

\item{e0}{threshold value for data maturity requirements C2}

\item{p0}{threshold value for data maturity requirements C3 (may takes a while is specified due to simulation)}

\item{m0}{threshold value for data maturity requirements C4}

\item{...}{the number of simulations and seed when p0 is specified. The default is nsim = 1000, seed = 12345. Change if needed}
}
\value{
Optimal set that maximize the financial benefits but at meantime, yields clinical meaningful results & data maturity constraints .
The results are saved in a list with three elements:
\item{OptRes}{The optimal trial design. A list including optimal (S, N, ra, Sa), but also Ea, actual power (AccPower), data maturity indicators (id.c1-4), total sale (Sale), total cost (Cost), total revenue (Rev), probability of observing clinical meaningful results (PA)}
\item{Designs}{A table of all potential trial designs}
\item{Valid.set}{A vector of indicators with the same length of dim(Designs)[1]. TRUEs indicate the designs satisfy the data maturity constraints}
}
\description{
A function to seek the optimal trial design in terms of both financial benefits and regulatory consideration, i.e. clinical meaningful results, and data maturity requirements
}
\details{
The function requires to input one of accrual rate (ra) and accrual period (Sa). If Sa is provided,
a uniform accrual rate will be assumed.
Investigators may have a rough idea of the range of ra in general. Notice that under the assumptions, the larger ra always yield short study duration and thus a higher revenue, which is
more desired. Therefore, the optimal results must happen at max(ra). In reality, the accrual rate (ra) is always constrained by many limitations. So a natural choice is
to input max(ra) in the function, and obtain the optimal trial design from the ouput.
}
\examples{
findOpt(ra = NULL, Sa = 15, Ea=88, lambda=c(log(2)/20, log(2)/10),
eta=c(log(2)/120, log(2)/100), b=18, c0=10, c1=0.6, c2=0.1,
dL=180, Aj.Ind = "Median Ratio", d0_r0 = 1.4, t0 = 6)

# input ra, and more data constraints
findOpt(ra = c(10,20,30,40,36), Ea=88, lambda=c(log(2)/20, log(2)/10),
eta=c(log(2)/120, log(2)/100), b=18, c0=10, c1=0.6, c2=0.1,
dL=180, Aj.Ind = "Median Ratio", d0_r0 = 1.4, t0 = 10, e0 = 0.6, p0 = 0.7, m0 = 6)
}
\references{
(Nektar Therapeutics) An Optimal Design Strategy for Phase III Clinical Trials with Time-To-Event Endpoint
}
