% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_t.R
\name{H_t}
\alias{H_t}
\title{An ancillary function for E_piece1 & E_piece2 (Not for all)}
\usage{
H_t(lambda, eta, t, S)
}
\arguments{
\item{lambda}{Event rate, assuming an exponential distribution}

\item{eta}{Loss to follow-up rate, assuming an exponential distribution}

\item{t}{current time}

\item{S}{study duration}
}
\value{
A numeric value
}
\description{
Function to calculate the integration, H(t)
}
