% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.N.R
\name{inv.N}
\alias{inv.N}
\title{Calculate the time of accrual (Not for all)}
\usage{
inv.N(N, ra.seq, Acc)
}
\arguments{
\item{N}{Number of total samples. If sum(ra.seq)<N, it directly assumes using ra.seq[-1] (the last one) as the accrual rate for the rest samples}

\item{ra.seq}{Sequence of piecewise accrual rate.}

\item{Acc}{Number of already accrued samples}
}
\value{
Time of accural period
}
\description{
Function used for calculating the time of accrual time based on the given number of accrued samples. Used for piecewise accrual case.
}
